feat: organization settings organization specific username constraints (#10271)
# Which Problems Are Solved

Usernames have to be instance specific, only with a setting related to
the used organization domains it is possible to make them organization
scoped.

# How the Problems Are Solved

On the organization there are now additional settings, which serve the
purpose to set the usernames organization scoped, which guarantees that
they have to be only unique in the organization not the instance.

# Additional Changes

As the `userLoginMustBeDomain` setting sets the usernames unique in the
organization, this organization settings have to be combined with some
logic.

# Additional Context

Closes #10086