feat: ListUsersByMetadata protobuf (#10349)
# Which Problems Are Solved

There is a good amount of people from the community that need to
retrieve users matching some metadata stored on the user itself.

It is currently not possible to achieve that, the current solution
entails retrieving the user first, reading the metas and aggregating the
data together based on the metas.

We want to offer a way to search through user metas and return the user
info matching the search.

# How the Problems Are Solved

This PR contains only the protobuf so that it's easier to review.

# Additional Changes

Converter methods have been moved to their own package to avoid
cluttering and have better organization. The converter methods have also
been tested

# Additional Context

Discussion #9053