chore(tests): disable parallelism in integration tests
This attempts to increase consistency by disabling parallelism in the integration test.
I suspect with the increased amount of integration tests, there's a battle for resources on the test runner between the zitadel binary, test binary and postgreSQL server. This may cause projection handlers to fall behind, resulting in very eventual consistency, many retries in EventuallyWithT and test timeouts if consistency isn't reached within the provided timeout. Retries add to resource consumption, so it might actually make the problem worse as multiple tests are retrying on a loop in parallel.