docs: fix import in response manipulation example (#10503)
# Which Problems Are Solved

Fixes one of the two issues mentioned in:
https://github.com/zitadel/zitadel/issues/10502

The response manipulation example has the wrong import in the Golang
code.

https://zitadel.com/docs/guides/integrate/actions/testing-response-manipulation

# How the Problems Are Solved

Changes the example document to use the right import address.

# Additional Changes

None.

# Additional Context

Replace this example with links to related issues, discussions, discord
threads, or other sources with more context.
Use the Closing #issue syntax for issues that are resolved with this PR.
- Partially resolves #10502
- https://discord.com/channels/927474939156643850/1405813424901984380