chore: API with Nx (#10744)
# Which Problems Are Solved

We introduced Nx as our build and cache orchestrator.
However, we still have orchestration tasks in the Makefile and in GitHub
Actions workflows relating to the API.
To streamline the DevX accross the whole monorepo, this PR replaces all
these orchestrations with Nx targets.

# How the Problems Are Solved

- Makefile recipes are replaced with Nx targets.
- GitHub Actions workflows for PR checks replaced by a single nx
affected command in the check.yml workflow file.
- The Codecov action uses nx to build the API.

# Additional Context

- Depends on #10571