fix: validate IDP linking conditions
# Which Problems Are Solved

When auto-linking was enabled on an IdP, there was no check if linking to the found user is allowed, i.e. if the corresponding IdP is active in the user's organization or if external authentication in general was allowed.

# How the Problems Are Solved

- (Re)Check the login policy of the user's organization before linking the external identity.

# Additional Changes

None

# Additional Context

None

---------

Co-authored-by: Max Peintner <peintnerm@gmail.com>
(cherry picked from commit 33c51deb20402dd5720e32cfb0c1d5fdc752f2e0)